+++ translator/src/main/scala/soda/translator/extension/toscala/TranslationConstantToScala.scala
@@ -73,7 +73,7 @@ trait TranslationConstantToScala
 
   lazy val scala_extends_translation = "extends"
 
-  lazy val scala_entry_point = "object EntryPoint {\n  def main (args: Array [String]): Unit = Main ().main (args)\n}\n"
+  lazy val scala_entry_point = "object EntryPoint {\n  def main (args: Array [String]): Unit = Main_ ().main (args)\n}\n"
 
   /**
    * Scala 3 keywords: