+++ translator/src/test/resources/soda/translator/example/miniexample/MiniExample.scala
@@ -19,11 +19,11 @@ case class MiniExample_ ()
 }
 
 object EntryPoint {
-  def main (args: Array [String]): Unit = Main ().main (args)
+  def main (args: Array [String]): Unit = Main_ ().main (args)
 }
 
 
-trait MainClass
+trait Main
 {
 
   def main (arguments: Array [String]  ): Unit =
@@ -31,9 +31,9 @@ trait MainClass
 
 }
 
-case class Main ()
+case class Main_ ()
   extends
-    MainClass
+    Main
 {
 
 }